refactor(electrum): reduce scope to example-only one-liner sync
Reverts library API changes and consolidates one-liner sync logic into a single self-contained example in examples/example_electrum.

Changes:

- Revert public API additions in bdk_electrum

- Remove test_api_compatibility

- Remove separate example_one_liner crate

- Add ElectrumSync helper to examples/example_electrum/src/bin/one_liner_sync.rs
