+++ crates/electrum/Cargo.toml
@@ -21,9 +21,7 @@ electrum-client = { version = "0.24.0", features = [ "proxy" ], default-features
 [dev-dependencies]
 
 bdk_chain = { path = "../chain" }
-anyhow = "1.0"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
+
 
 criterion = { version = "0.7" }
