+++ crates/electrum/src/lib.rs
@@ -24,5 +24,4 @@ pub use bdk_electrum_client::*;
 
 
 
-pub use bdk_core;
-pub use electrum_client;
+
