+++ examples/example_electrum/src/bin/one_liner_sync.rs
@@ -1,22 +1,126 @@
-//! Example of a one-liner wallet sync using ElectrumSync.
-//!
-//! This example demonstrates how to:
-//! 1. Create a wallet (KeychainTxOutIndex).
-//! 2. Create an Electrum client.
-//! 3. Use `ElectrumSync` for a "one-liner" sync.
-//!
-//! Note: This example requires an actual Electrum server URL to run successfully.
-//! By default it tries to connect to a public testnet server.
-
 use bdk_chain::{
-    bitcoin::{Network, Network::Testnet},
-    keychain_txout::KeychainTxOutIndex, // Correct import path
+    bitcoin::Network,
+    keychain_txout::KeychainTxOutIndex,
+    tx_graph::TxGraph,
+    collections::BTreeMap,
+    CheckPoint,
 };
-use bdk_electrum::{
-    electrum_client::{self, ElectrumApi},
-    BdkElectrumClient, ElectrumSync, SyncOptions,
+use bdk_core::{
+    spk_client::{FullScanRequest, SyncRequest},
+    ConfirmationBlockTime,
 };
-use std::collections::BTreeMap;
+use bdk_electrum::BdkElectrumClient;
+use electrum_client::{self, ElectrumApi};
+
+// -----------------------------------------------------------------------------
+// ONE-LINER SYNC HELPER (Proposed API Pattern)
+// -----------------------------------------------------------------------------
+
+// NOTE: This helper is intentionally defined in the example.
+// It demonstrates how an application may compose existing BDK APIs.
+// This is NOT part of the bdk_electrum public API.
+
+/// Configuration for the sync operation.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct SyncOptions {
+    pub fast: bool,
+    pub stop_gap: usize,
+    pub batch_size: usize,
+    pub fetch_prev: bool,
+}
+
+impl Default for SyncOptions {
+    fn default() -> Self {
+        Self {
+            fast: true,
+            stop_gap: 20,
+            batch_size: 10,
+            fetch_prev: false,
+        }
+    }
+}
+
+impl SyncOptions {
+    pub fn fast() -> Self {
+        Self { fast: true, ..Default::default() }
+    }
+    
+    pub fn full_scan() -> Self {
+        Self { fast: false, ..Default::default() }
+    }
+}
+
+pub struct ElectrumSync<'a, K, E> {
+    wallet: &'a KeychainTxOutIndex<K>,
+    client: BdkElectrumClient<E>,
+}
+
+impl<'a, K, E> ElectrumSync<'a, K, E>
+where
+    E: ElectrumApi,
+    K: Ord + Clone + core::fmt::Debug + Send + Sync,
+{
+    pub fn new(wallet: &'a KeychainTxOutIndex<K>, client: BdkElectrumClient<E>) -> Self {
+        Self { wallet, client }
+    }
+
+    pub fn sync(
+        &self,
+        options: SyncOptions,
+    ) -> Result<
+        (
+            Option<CheckPoint>,
+            TxGraph<ConfirmationBlockTime>,
+            Option<BTreeMap<K, u32>>,
+        ),
+        electrum_client::Error,
+    > {
+        if options.fast {
+            let request = SyncRequest::builder()
+                .spks_with_indexes(
+                    self.wallet
+                        .revealed_spks(..)
+                        .map(|(k, spk)| (k.1, spk.into())),
+                )
+                .build();
+
+            let response = self
+                .client
+                .sync(request, options.batch_size, options.fetch_prev)?;
+
+            Ok((
+                response.chain_update,
+                response.tx_update.into(),
+                None,
+            ))
+        } else {
+            let mut builder = FullScanRequest::builder();
+            
+            for (keychain, spks) in self.wallet.all_unbounded_spk_iters() {
+                builder = builder.spks_for_keychain(keychain, spks);
+            }
+
+            let request = builder.build();
+
+            let response = self.client.full_scan(
+                request,
+                options.stop_gap,
+                options.batch_size,
+                options.fetch_prev,
+            )?;
+
+            Ok((
+                response.chain_update,
+                response.tx_update.into(),
+                Some(response.last_active_indices),
+            ))
+        }
+    }
+}
+
+// -----------------------------------------------------------------------------
+// EXAMPLE USAGE
+// -----------------------------------------------------------------------------
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 enum MyKeychain {
@@ -27,21 +131,12 @@ enum MyKeychain {
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     const ELECTRUM_URL: &str = "ssl://electrum.blockstream.info:60002"; // Testnet
 
-    // 1. Setup Wallet: KeychainTxOutIndex
     let mut wallet_index = KeychainTxOutIndex::<MyKeychain>::new(20, true);
-
-    // Add descriptors (using some public descriptor for demo purposes)
-    // Descriptor: tr([73c5da0a/86'/1'/0']tpubDC.../0/*) (External)
-    // This is just a dummy descriptor for compilation, won't find real funds on testnet unless the xpub is valid/funded
-    let external_descriptor = "tr([73c5da0a/86'/1'/0']tpubDCDkM3bAi3d7KqW8G9w8V9w8V9w8V9w8V9w8V9w8V9w8V9w8V9w8V9w8V9w8V9w8V-testnet/0/*)";
-    // Note: Parsing descriptors requires more boilerplate in real code (miniscript, secp256k1), 
-    // omitted here for brevity if just checking API structure. 
-    // BUT we need it to compile. So let's use a simpler known descriptor if possible, or just mock usage.
-    
-    // For the sake of this example being purely about API structure, we will skip actual descriptor parsing
-    // unless we need to query specifically. In a real app you'd insert descriptors here.
     println!("Wallet index initialized.");
 
+    // This descriptor is specific to Testnet.
+    // In a real example we might parse it, but for now we just initialize index.
+    
     // 2. Setup Electrum Client
     let electrum_client = electrum_client::Client::new(ELECTRUM_URL)?;
     // Wrap it in BdkElectrumClient (preserves cache)
@@ -57,8 +152,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let result = syncer.sync(SyncOptions::full_scan())?;
     
     // Ideally we would apply the result to the wallet here.
-    // wallet_index.apply_changeset(result.2.unwrap()); // Applying last active indices
-    // wallet_index.apply_update(result.1); // Applying tx graph
+    // wallet_index.apply_update(result.1, result.2); // Conceptual, depends on specific Wallet/Index API for application.
     
     println!("Sync finished!");
     println!("New tip: {:?}", result.0);
@@ -67,7 +161,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     // 4. Repeated Sync (Fast)
     // Suppose we just want to check revealed addresses for new txs (faster).
     println!("Starting fast sync...");
-    let fast_result = syncer.sync(SyncOptions::fast())?;
+    let _fast_result = syncer.sync(SyncOptions::fast())?;
     
     println!("Fast sync finished!");
