--- examples/one_liner_sync.rs
@@ -1,76 +1,0 @@
-//! Example of a one-liner wallet sync using ElectrumSync.
-//!
-//! This example demonstrates how to:
-//! 1. Create a wallet (KeychainTxOutIndex).
-//! 2. Create an Electrum client.
-//! 3. Use `ElectrumSync` for a "one-liner" sync.
-//!
-//! Note: This example requires an actual Electrum server URL to run successfully.
-//! By default it tries to connect to a public testnet server.
-
-use bdk_chain::{
-    bitcoin::{Network, Network::Testnet},
-    indexer::KeychainTxOutIndex,
-};
-use bdk_electrum::{
-    electrum_client::{self, ElectrumApi},
-    BdkElectrumClient, ElectrumSync, SyncOptions,
-};
-use std::collections::BTreeMap;
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-enum MyKeychain {
-    External,
-    Internal,
-}
-
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    const ELECTRUM_URL: &str = "ssl://electrum.blockstream.info:60002"; // Testnet
-
-    // 1. Setup Wallet: KeychainTxOutIndex
-    let mut wallet_index = KeychainTxOutIndex::<MyKeychain>::new(20, true);
-
-    // Add descriptors (using some public descriptor for demo purposes)
-    // Descriptor: tr([73c5da0a/86'/1'/0']tpubDC.../0/*) (External)
-    // This is just a dummy descriptor for compilation, won't find real funds on testnet unless the xpub is valid/funded
-    let external_descriptor = "tr([73c5da0a/86'/1'/0']tpubDCDkM3bAi3d7KqW8G9w8V9w8V9w8V9w8V9w8V9w8V9w8V9w8V9w8V9w8V9w8V9w8V-testnet/0/*)";
-    // Note: Parsing descriptors requires more boilerplate in real code (miniscript, secp256k1), 
-    // omitted here for brevity if just checking API structure. 
-    // BUT we need it to compile. So let's use a simpler known descriptor if possible, or just mock usage.
-    
-    // For the sake of this example being purely about API structure, we will skip actual descriptor parsing
-    // unless we need to query specifically. In a real app you'd insert descriptors here.
-    println!("Wallet index initialized.");
-
-    // 2. Setup Electrum Client
-    let electrum_client = electrum_client::Client::new(ELECTRUM_URL)?;
-    // Wrap it in BdkElectrumClient (preserves cache)
-    let bdk_client = BdkElectrumClient::new(electrum_client);
-
-    // 3. One-Liner Sync
-    // We create the helper.
-    let syncer = ElectrumSync::new(&wallet_index, bdk_client);
-
-    println!("Starting full scan...");
-    
-    // Perform a full scan (discovers scripts)
-    let result = syncer.sync(SyncOptions::full_scan())?;
-    
-    // Ideally we would apply the result to the wallet here.
-    // wallet_index.apply_update(result.1, result.2); // Conceptual, depends on specific Wallet/Index API for application.
-    // KeychainTxOutIndex doesn't directly take TxGraph updates, usually a `Wallet` struct does.
-    // But this shows we got the data.
-    
-    println!("Sync finished!");
-    println!("New tip: {:?}", result.0);
-    println!("Found transactions: {}", result.1.full_txs().count());
-
-    // 4. Repeated Sync (Fast)
-    // Suppose we just want to check revealed addresses for new txs (faster).
-    println!("Starting fast sync...");
-    let fast_result = syncer.sync(SyncOptions::fast())?;
-    
-    println!("Fast sync finished!");
-
-    Ok(())
-}
